+++ ktlint/src/test/kotlin/com/pinterest/ktlint/BaseCLITest.kt
@@ -9,6 +9,7 @@ import java.nio.file.SimpleFileVisitor
 import java.nio.file.attribute.BasicFileAttributes
 import java.util.concurrent.TimeUnit
 import org.assertj.core.api.Assertions.assertThat
+import org.junit.jupiter.api.fail
 import org.junit.jupiter.api.io.TempDir
 
 abstract class BaseCLITest {
@@ -34,21 +35,42 @@ abstract class BaseCLITest {
                 it.replace(BASE_DIR_PLACEHOLDER, tempDir.toString())
             }
         // Forking in a new shell process, so 'ktlint' will pickup new 'PATH' env variable value
-        val pb = ProcessBuilder("/bin/sh", "-c", ktlintCommand)
-        pb.directory(projectPath.toAbsolutePath().toFile())
+        val processBuilder = ProcessBuilder("/bin/sh", "-c", ktlintCommand)
+        processBuilder.directory(projectPath.toAbsolutePath().toFile())
 
         // Overriding user path to java executable to use java version test is running on
-        val environment = pb.environment()
+        val environment = processBuilder.environment()
         environment["PATH"] = "${System.getProperty("java.home")}${File.separator}bin${File.pathSeparator}${System.getenv()["PATH"]}"
 
-        val process = pb.start()
-        val output = process.inputStream.bufferedReader().use { it.readLines() }
-        val error = process.errorStream.bufferedReader().use { it.readLines() }
-        process.waitFor(WAIT_TIME_SEC, TimeUnit.SECONDS)
+        val process = processBuilder.start()
+        if (process.completedInAllowedDuration()) {
+            val output = process.inputStream.bufferedReader().use { it.readLines() }
+            val error = process.errorStream.bufferedReader().use { it.readLines() }
 
-        executionAssertions(ExecutionResult(process.exitValue(), output, error, projectPath))
+            executionAssertions(ExecutionResult(process.exitValue(), output, error, projectPath))
 
-        process.destroy()
+            // Destroy process only after output is collected as other the streams are not completed.
+            process.destroy()
+        } else {
+            // Destroy before failing the test as the process otherwise keeps running
+            process.destroyForcibly()
+
+            val maxDurationInSeconds = (WAIT_INTERVAL_DURATION * WAIT_INTERVAL_MAX_OCCURRENCES).div(1000.0)
+            fail {
+                "CLI test has been aborted as it could not be completed in $maxDurationInSeconds seconds"
+            }
+        }
+    }
+
+    private fun Process.completedInAllowedDuration(): Boolean {
+        (0..WAIT_INTERVAL_MAX_OCCURRENCES).forEach { _ ->
+            if (isAlive) {
+                waitFor(WAIT_INTERVAL_DURATION, TimeUnit.MILLISECONDS)
+            } else {
+                return true
+            }
+        }
+        return false
     }
 
     private fun prepareTestProject(testProjectName: String): Path {
@@ -128,7 +150,8 @@ abstract class BaseCLITest {
     }
 
     companion object {
-        private const val WAIT_TIME_SEC = 3L
+        private const val WAIT_INTERVAL_DURATION = 100L
+        private const val WAIT_INTERVAL_MAX_OCCURRENCES = 100
         val testProjectsPath: Path = Paths.get("src", "test", "resources", "cli")
         const val BASE_DIR_PLACEHOLDER = "__TEMP_DIR__"
     }